+++ src/main/java/kr/co/pawong/pwbe/test/adapter/in/messaging/kafka/KafkaTestConsumer.java
@@ -1,7 +1,7 @@
 package kr.co.pawong.pwbe.test.adapter.in.messaging.kafka;
 
 import kr.co.pawong.pwbe.test.application.port.in.PrintLogUseCase;
-import kr.co.pawong.pwbe.test.config.KafkaTopicConfig;
+import kr.co.pawong.pwbe.global.config.KafkaTopicConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
