+++ src/main/java/kr/co/pawong/pwbe/test/application/service/PublishTestMessageService.java
@@ -2,7 +2,7 @@ package kr.co.pawong.pwbe.test.application.service;
 
 import kr.co.pawong.pwbe.test.application.port.in.PublishTestMessageUseCase;
 import kr.co.pawong.pwbe.test.application.port.out.TestMessagePublishPort;
-import kr.co.pawong.pwbe.test.config.KafkaTopicConfig;
+import kr.co.pawong.pwbe.global.config.KafkaTopicConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
